Fix for default build locations
Previously, if multiple sites did not have a destination path specified, they would all go into the build_local folder. This is fine for a single install, but with multiple sites, each would progressively overwrite the previous. This change defaults the build locations for sites to be build_local/site_name, only if there is a "sites" array in the root config file.

A long comment is also included to explain this.