Adds support to inline asset (css/js)
depending on the size of your CSS or JS, you may want to inline them
into the style or script tag of the head of your pages so you save
one request and avoid blocking the rest of the page from loading.

Due to that, we added the `inline` function that will be used in
conjunction to the `mix` function to properly get the content
of your asset file and print it where we want it.